system_power: Document power states and helper preconditions
Because there is no synchronization in the system_state_machine() loop,
it is not safe to call the state transition helper functions while a
transition is in progress. While this could be made to work, it is not
necessary at the moment. Transition helper functions are only currently
(and expected to be) called from polling functions. This means scpi_poll
or equivalent during ACTIVE, and a future wakeup device-polling function
during INACTIVE.

Signed-off-by: Samuel Holland <samuel@sholland.org>