system: Only run the watchdog while the system is active
Save that extra little bit of power by turning off the watchdog while
the firmware is idle (spinning in a tight loop waiting for an IRQ). By
disabling it last and enabling it first, we minimize the unprotected
window.

Since a reset can happen at any time, that code path must work whether
or not there is already a reference to the watchdog.

Signed-off-by: Samuel Holland <samuel@sholland.org>