css: Fix state inspection functions
ACINACTM is not a reliable indicator of the cluster's reset status. It
should be "1" when the cluster is in reset... except that the cluster is
turned off and not producing outputs. Let's use an input to the cluster
(one that has a stable meaning and location), not an output from it.

Fixes: 3a7b72427d9 ("css: a64: Simplify state inspection functions")
Signed-off-by: Samuel Holland <samuel@sholland.org>