main: Use the right macro for the EPCR SPR
SPR_SYS_EPCR_INDEX gets the index from the start of the SPR group, while
SPR_SYS_EPCR_ADDR gets the full address. They happen to be
interchangeable here because EPCR is in the first SPR group, but only
the semantic meaning of SPR_SYS_EPCR_ADDR is correct for the mtspr()
wrapper function, which expects the full zero-based address.

Fixes: b50533d5cd48 ("common: Reset the firmware after an exception")
Signed-off-by: Samuel Holland <samuel@sholland.org>