drivers: css: Silence unused variable warnings
Since there is only one cluster on A64/H5/H6, the cluster parameter is
not needed to query cluster/core state. Usually, the assertion causes
the parameter to be referenced, but when assertions are disabled, a
warning is generated. Mark the parameter as possibly unused, so there is
no warning.

Signed-off-by: Samuel Holland <samuel@sholland.org>