travis: Disable config options that cause overflow
Now that there are debugging options that add a significant amount of
code/data to the firmware, an `allyesconfig` or `randconfig` build may
overflow the area available in SRAM A2. This indicates an invalid
configuration, not a problem with the code.

Keep such options disabled during CI builds for reliability.

Signed-off-by: Samuel Holland <samuel@sholland.org>