system: Ensure R_PIO's clock is always ungated at runtime
Linux cannot currently handle resuming with R_PIO's clock gated.
Specifically, if R_PIO is used as a wakeup source, Linux will not gate
the clock during suspend. This is required for the irqchip to function
during s2idle and detect pin transitions. And to keep enable/disable
balanced, it will not ungate clock during resume in that case either.

So if a Port L GPIO IRQ causes wakeup, Linux will assume that the clock
is already running. If that happens to be false, Linux will be unable to
clear the IRQ, which will cause a hard lockup.

Similar to the watchdog, we take a reference to R_PIO any time Linux is
running. We also keep the reference during suspend, if R_PIO is
configured as a wakeup source.

Signed-off-by: Samuel Holland <samuel@sholland.org>