drivers: rsb: Hardcode the SRTA and PMCR constants
These don't appear to have ever changed, and are hardcoded in the Linux
driver as well. Since the PMCR values are the same for all RSB PMICs, we
can do the mode change once for the whole bus.

Linux has a lookup table for hwaddr<->rtaddr, but as the bus driver only
needs to support a single (primary) PMIC at the moment, use an assertion
here instead.

This removes the extra arguments from the probe function, allowing a
future merge with the corresponding i2c probe function.

Signed-off-by: Samuel Holland <samuel@sholland.org>