drivers: Rename "rsb" to "regmap" to prepare for merge with I2C
In order to support PMICs connected via either RSB or I2C, the RSB and
I2C classes need to have a common set of types and functions. This is
basically the regmap interface, as implemented in Linux and U-Boot.
Since RSB already has the desired regmap-like API at the driver ops
level, use it as the model for the common class. I2C drivers will need
another layer of abstraction to fit this interface.

Signed-off-by: Samuel Holland <samuel@sholland.org>