drivers: Convert I2C drivers to the regmap class
This works by using the generic I2C wrapper functions to implement the
regmap driver ops as "regmap-i2c". They then call into a second set of
lower-level driver ops, as before. So the driver is largely unchanged.

Now a driver for an external chip doesn't need to care if it is
connected via I2C or RSB.

Signed-off-by: Samuel Holland <samuel@sholland.org>