drivers: regulator: axp803: Clean up some messy logic
Instead of having a mask of bits to ignore, specify the mask of actual
value bits in the register. Then we can reuse that mask for writing with
regmap_update_bits. And instead of doing some awful xor cleverness (that
was wrong, because it only looked at one bit of the GPIO pin function),
use some better-readable if statements to handle the GPIO special cases.

Signed-off-by: Samuel Holland <samuel@sholland.org>