device: Gracefully handle acquiring the NULL device
When passing handles around, it is useful describe an empty handle with
a NULL device. Since anything that uses a handle must acquire the
providing device before using it, this is safe as long as device_get()
returns an error for a NULL device. Then it will be treated the same as
if the device had failed to probe for any other reason.

Signed-off-by: Samuel Holland <samuel@sholland.org>