system: Disable VDD-CPUX in inactive/off states
This regulator should be safe to disable on every board where it is
present, as crust is responsible for disabling all of its consumers.

It is necessary to re-enable regulators before watchdog reset, as that
only resets the SoC, not any external components. CPU power must be
enabled for the SoC to reset correctly.

Since all regulators are expected to take some time to turn back on,
do it once at the end, instead of in the individual driver.

Signed-off-by: Samuel Holland <samuel@sholland.org>