drivers: regulator: Remove value getting/setting
The firmware does not actually care about the voltage output level of
regulators. The initial regulator setup is done by ATF or U-Boot SPL,
and regulators are expected to remember their voltage during a suspend/
resume cycle (when they are disabled/re-enabled). Even if they don't,
Linux will fix the voltage during resume -- if the default voltage isn't
close enough to allow Linux to resume, it wouldn't be close enough for
Linux to boot in the first place.

Signed-off-by: Samuel Holland <samuel@sholland.org>