drivers: ccu: Don't hardcode register offsets in hooks
The goal of the hook-based architecture is to share code between clocks
that work the same way. When a clock is unique, and there's only one
relevant register, hardcoding it is a win. But as soon as the helper
needs to be shared among more than two clocks, it becomes simpler to
read the register from the clock's data than trying to ternary/switch it
from the ID.

Signed-off-by: Samuel Holland <samuel@sholland.org>