debug: Allow COMPILE_TEST to fill the full address space
Removing the assertions only allows the allocation for the stack to run
past the end of SRAM A2. ld(1) produces an error if it has to allocate
an object outside the memory region we defined. So to remove *all* size
restrictions, the memory region must also be enlarged to the size of the
address space.

Signed-off-by: Samuel Holland <samuel@sholland.org>