scpi: Directly use protocol-defined error codes
In order to distinguish error paths with minimal code size, we can use
the address of the original error return as the return code. However,
this breaks scpi_map_error.

Since the SCPI commands are relatively isolated from the rest of the
codebase, we can use the protocol-defined error constants directly,
instead of mapping to them. The CSS drivers already use SCPI protocol
constants, so this adds no additional coupling.

Signed-off-by: Samuel Holland <samuel@sholland.org>