drivers: ccu: Support clocks with update bits
Some clock registers ignore changes until a specific bit is set. When
the clock gate is in one of these registers, we need to pulse this
update bit after enabling/disabling the gate. This is the case for at
least PLL_DDR0/1 and the DRAM module clock on the A64.

Signed-off-by: Samuel Holland <samuel@sholland.org>