drivers: sun50i-a64-css: Implement DRAM self-refresh
The DRAM controller suspend procedure generally follows the BSP
procedure. The main difference in this version is that it does not
assert the DRAM controller APB domain reset; that definition was
explicitly omitted from the CCU driver. This allows the resume procedure
to simply mirror and undo the suspend procedure, instead of needing to
save several registers and reinitialize the DRAM controller.

The names for the registers and bitfields generally come from the
Synopsys documentation (with pieces filled in from the UltraScale+ and
Keystone II documentation) after determining their meaning from the BSP
and mainline U-Boot code. They are cross-referenced to their usage in
the BSP code for future analysis.

Signed-off-by: Samuel Holland <samuel@sholland.org>