Makefile: Separate BUILD from HOST toolchain
This project creates executables for three different machines:
  1) the build machine, where compilation is done (Kconfig)
  2) the host machine, the Linux environment on a sunxi SoC (tools)
  3) the target machine, the or1k coprocessor on a sunxi SoC (scp.bin)

For that reason, we need three separate toolchains.

The existing meaning of CROSS_COMPILE is kept for compatibility, and a
new HOST_COMPILE variable is created, which is used to provide the
prefix for the HOST cross toolchain. In the event that the firmware is
built on an arm/aarch64 system, HOST_COMPILE can be set to an empty
string.

This adds infrastructure to Makefile.kbuild to handle a new "buildprogs"
verb, which reuses the "hostprogs" implementation, but replaces the
HOST* variables (e.g. HOSTCC, HOSTCFLAGS) with the equivalent BUILD*
variables. This new "buildprogs" is used for compiling Kconfig.

Signed-off-by: Samuel Holland <samuel@sholland.org>