project: Move arch-specific files to a new directory
This solves the longstanding issue where "spr.h" was just as
architecture-specific as "arch/spr.h", but not marked as such.
"arch/spr.h" was really the assembly-safe subset of "spr.h", so
it should go in a directory named accordingly. (Back when there
was an "exception.h", it had the same problem.)

This also neatly solves the issue where files in lib/ could not
actually go in lib.a, because they were architecture-specific.

Signed-off-by: Samuel Holland <samuel@sholland.org>