drivers: mfd: Use a single option for the PMIC driver
Previously it was not possible to exclude the AXP20x MFD driver when
enabling the debug monitor. And there's no real reason to enable one of
the PMIC or regulator subdevice drivers, but not the other. Either the
chip is present or it is not. So create a single option to consistently
enable or disable the whole thing.

Signed-off-by: Samuel Holland <samuel@sholland.org>