device: Split part of device_get into device_get_or_null
There are two general use cases for device_get: either the caller
controls the lifetime of the device, and they want a pointer, or the
caller is working with an existing reference to a device (such as in a
handle), and they want to pass back a possible error from the device
probe.

Satisfy both use cases by splitting the function into two.

Signed-off-by: Samuel Holland <samuel@sholland.org>