device: Gracefully handle releasing the NULL device
Commit f03b906d5553 ("device: Gracefully handle acquiring the NULL
device") improved the acquire-side function. This commit does the same
for the release side.

Accepting a NULL pointer means that, as long as the reference is cleared
after putting the device, we can keep calling device_put() without
needing a NULL check. This removes quite a few branches in favor of a
single branch in device_put().

Signed-off-by: Samuel Holland <samuel@sholland.org>