drivers: watchdog: Implicitly enable watchdogs when probing
There is not much use to a probed but disabled watchdog. This makes
watchdog probe/release symmetric -- watchdogs were already disabled when
releasing.

This was not done earlier because there is no way to pass a timeout
parameter to the probe function. However, it works to start with a
hardcoded conservative timeout. The timeout can be reduced later if
needed, such as when intentionally triggering a restart.

Signed-off-by: Samuel Holland <samuel@sholland.org>