common: Introduce the simple_device abstraction
This abstraction represents an MMIO device with a clock and optionally a
pair of pins. This construction meets the needs of most SoC-internal
devices used by Crust so far, as well as soon-to-be-added devices, like
UARTs.

The purpose of this abstraction is not really to save space; its goal is
to allow code outside the individual drivers to check the clock status
for a device. The clock status can be used as a proxy for if Linux has
implemented/enabled device PM for a device. Checking the clock status is
only efficient if all of the relevant devices share a way of accessing
their clock handle (i.e. use the same struct).

Signed-off-by: Samuel Holland <samuel@sholland.org>