debug: Merge the PANIC log level into ERROR
The firmware is designed to immedately restart when encountering an
unhandled exception (like an explicit trap). And the firmware is
designed to safely restart from any system power state. So there is no
unrecoverable error where the firmware must give up and spin in a loop,
and there's no distinction between a PANIC and an ERROR.

Signed-off-by: Samuel Holland <samuel@sholland.org>