device: Add a function to check device status
A similar function was previously removed for breaking abstraction.
Ideally, one should never pass an unreferenced device pointer to
anything but device_get[_or_null](). However, for code that can be
called during device probe from a non-driver context (e.g. log()),
it is necessary to have a global-variable view of a device's state.
In that case, saving the return value of device_get() in a global
variable is very wasteful -- the refcount already provides a global
view of if a device is available to interact with. Now generally
this would be unsafe, because whatever has that reference could drop
it. But in the specific case of a device with at least one persistent
reference, looking at the refcount directly is a useful optimization.

Signed-off-by: Samuel Holland <samuel@sholland.org>