common: Move serial init after CCU init
For further power saving, CLK_AR100 needs to use its OSC16M parent.
However, changing the AR100 clock parent only works if there are no
references to it, as otherwise the reference counts on the oscillators
would be wrong. So R_UART init must come after CCU init.

This provides an opportunity to move the arch-specific exception
reporting code to the correct directory.

Signed-off-by: Samuel Holland <samuel@sholland.org>