drivers: ccu: Reparent CLK_AR100 to OSC16M
This is necessary to disable OSC24M during suspend.

Because the frequency of OSC16M is highly variable between chips, it
must be calibrated (measured in comparison to OSC24M) to get a frequency
accurate enough for UART communication. Even then, the UART symbol rate
must be reduced by an order of magnitude to work reliably.

Signed-off-by: Samuel Holland <samuel@sholland.org>