drivers: serial: Always consider preinitialized UARTs ready
If the UART is assumed to be initialized before the firmware starts, we
can start using it immediately. The probe function should be a no-op in
this case.

Note that we still need to re-enable the clock in case it is later
disabled (e.g. by Linux's clk_disable_unused), even though the clock is
known to be enabled at boot.

Signed-off-by: Samuel Holland <samuel@sholland.org>