debug: Fix issues with latency measurement
Reducing the bus clock frequencies to tens of kHz significantly
increases the time needed to perform one iteration. Reduce the number of
iterations by a couple of orders of magnitude, so the measurement is
printed after a reasonable time, and so the math does not overflow.

Clean up the condition which caused the denominator to be off by one.

Use unsigned formatting conversions for unsigned numbers.

Signed-off-by: Samuel Holland <samuel@sholland.org>