drivers: sun8i-r-ccu: Only set VDD_SYS gating on A64/H5
The relevant VDD_SYS_PWROFF_GATING_REG bits are undocumented, and the
register layout is known to vary between SoCs. To be safe, we should
only set bits on the specific platforms where they have been tested.

On the A64 and H5 SoCs, setting bit 2 is known to be safe and improve
power consumption. It is likely a PLL gating bit, but its exact function
has not been confirmed.

Signed-off-by: Samuel Holland <samuel@sholland.org>