drivers: ccu: Improve comments for simplifications
While working on adding support for the H3 CCU, I ran across some
questionable code. In all cases, the code was working as intended; I had
intentionally deviated from accurately describing the hardware to
simplify the driver. To avoid confusion, such places should be commented
so they are not seen as bugs.

Signed-off-by: Samuel Holland <samuel@sholland.org>