drivers: ccu: Hide RSB clock behind HAVE_RSB
While the R_CCU layout is stable across SoC generations, not all
peripherals are present in every chip. For example, the RSB controller
is not present on H-series chips, like the H3. Since the platforms
having an RSB controller are already enumerated for the HAVE_RSB config
item, use that config to avoid enumerating them again here.

Signed-off-by: Samuel Holland <samuel@sholland.org>