drivers: regmap: Provide I2C pin config via Kconfig
The logic for choosing the correct pin mode for PL0 and PL1 is more
complicated than will fit on a single line. The H3 is pin-compatible
with the H5, and it also requires a mode of 2. A more general rule is
that only chips with an RSB controller put I2C on mode 3 (which makes
sense, as RSB uses mode 2). Then H6 becomes the exception, not H5.

Signed-off-by: Samuel Holland <samuel@sholland.org>