scripts: Always clean up temporary header
If the existing header matched the temporary header, it would not be
replaced; and since the trap is cleared, the temporary header would be
left around. This is wrong. In fact, there is no case where running `rm`
is worse than not running it (even if `mv` succeeds), so always leave
the trap in place.

Signed-off-by: Samuel Holland <samuel@sholland.org>