arch: or1k: COMPILE_TEST should not move SCPI shmem
One purpose of COMPILE_TEST is to turn off the firmware size checks so
allyesconfig and randconfig builds can succeed. However, given an
otherwise functioning config (such as a board defconfig), enabling
COMPILE_TEST should by itself have no effect on the compiled firmware.

Previously, this rule was broken, and so was the resulting firmware
binary, as the method used to disable the size checks also moved the
SCPI shared memory area. This commit changes the linker script to
disable the size checks in a way that does not move any memory regions.

Signed-off-by: Samuel Holland <samuel@sholland.org>