drivers: sun6i-i2c: Use error codes, not Booleans
It is inefficient to generate a Boolean result for comparison against
zero, and then later convert that Boolean into an integer error code.

Instead, generate the error code from the start, so it can be used
unmodified further up the call stack. This saves about 48 bytes.

While this changes one error in sun6i_i2c_start from ENODEV to EIO, the
initial call to this function is from regmap_i2c_prepare, which converts
all errors to ENODEV anyway. So the change is only seen if the device
responds at first and then later fails to respond; EIO seems appropriate
for that condition.

Signed-off-by: Samuel Holland <samuel@sholland.org>