arch: or1k: Optimize pointer loads in startup code
Since the linker script guarantees the high word of all symbol addresses
is the same, it only needs to be loaded once. Since the symbol argument
to hi() does not matter, use the current function.

Now that initializing the stack pointer only takes one instruction
again, that code can be moved back to its original location.

Signed-off-by: Samuel Holland <samuel@sholland.org>