arch: or1k: Process arguments after enabling icache
Other than the magic instruction at the very beginning of the binary,
the first thing the firmware does should be to enable the instruction
cache. Otherwise, some code will be twice as slow for no benefit.

Moving the argument processing to just before the function call also
saves an instruction by computing the value directly in r3.

Signed-off-by: Samuel Holland <samuel@sholland.org>