arch: or1k: Split startup assembly into sections
The existing assembly is smashed together, making it hard to read, and
block comments are interspersed with line comments. Move the block
comments to their own lines to make the code easier to parse.

The duplicate line comments will be removed in the next commit. Moving
the comments in one commit makes the diff extremely confusing to read.

Signed-off-by: Samuel Holland <samuel@sholland.org>