system: Persist the system state across restarts
This allows using the actual state during startup, instead of guessing a
state based on the set of online CPUs. Not only is this much more
efficient, but is is more reliable: the state-guessing would have failed
for a multi-cluster SoC like the A83T if the system was active but all
cores in the first cluster were offline.

Signed-off-by: Samuel Holland <samuel@sholland.org>