arch: or1k: Relax the restriction on persistent data
Some data, such as the calibrated OSC16M frequency and the system state,
really do need to be persisted across firmware exception restarts.
Currently, this is accomplished by lying to the compiler (for OSC16M) or
attempting to reconstruct the value after the fact (system state).

I implemented a "persistent .bss" area, with logic to clear the area if
the firmware was reloaded (by a system reset)... just to realize I was
reimplementing the semantics of .data. The downside of using .data
directly is that it's easy to accidentally put things in by forgetting
to use "const". For that reason, I'm keeping the assertion, so adding
things to .data requires explicit manual intervention.

Signed-off-by: Samuel Holland <samuel@sholland.org>