drivers: sunxi-rsb: Support arbitrary hardware addresses
Currently, the code supports a single hardware address (the one used by
the AXP803 PMIC) but arbitrary runtime addresses. This is backward!
Since the runtime address is assignable, we only need to support as many
runtime addresses as we support simultaneous independent devices. Right
now, that's only one: an AXP20x PMIC (or an AXP20x master/slave pair).

So instead of hardcoding the hardware address 0x3a3, hardcode the
runtime address 0x2d.

Putting the hardware address in the handle allows supporting PMIC
variants with arbitrary hardware addresses, for example the AXP805.
Doing so requires expanding the `id` field in the handle, but the
structure was already padded out to 8 bytes anyway.

Signed-off-by: Samuel Holland <samuel@sholland.org>