platform: Add/use platform PRCM register definitions
The basic logic for controlling the PRCM is the same across sunxi SoC
generations, but the registers have moved around. Furthermore, the PRCM
registers are only sparsely documented, with small bits of information
gathered from each of several sources.

Document known information about the PRCM registers and bits so they can
be compared across SoC generations. Put the information in a platform-
specific header, so it can be shared across all drivers that need it,
and so the drivers themselves can be maximally reusable across SoCs.

Signed-off-by: Samuel Holland <samuel@sholland.org>