drivers: ccu: Factor out common PRCM driver logic
Now that the PRCM registers and fields have been abstracted by a
platform header, and the register fields are better understood to match
among hardware generations, the "non-CCU" part of the PRCM driver code
can be shared between drivers.

In addition to moving code around, this commit enables VCC_PLL_GATING on
H6 (where it has been tested) and A83T (where it has not).

Signed-off-by: Samuel Holland <samuel@sholland.org>