drivers: ccu: Improve code that writes PLL_CTRL_REG1
PLL_CTRL_REG1 is one of several registers with a "key" field. If the
current value in the key field does not match a magic value, only the
key field can be written. If it does match, the whole register can be
written. In either case, the value written to the key field is not used
until the _next_ write cycle.

So if the register is already unlocked, the first write will take
effect (thus it must be a good value, not just the key). Otherwise, only
the second write will take effect. Both cases can be handled without a
branch by writing the same value twice.

On the other hand, the register can be updated and locked in a single
write cycle.

The function is also updated to write the LDO and crystal enable bits
separately, with a delay take from the vendor blob.

Signed-off-by: Samuel Holland <samuel@sholland.org>