drivers: regmap: Hook up RSB on the H6 platform
While it is not documented, an RSB controller exists inside the H6, and
is found to be fully functional. Keep it disabled by default, as
switching the pinmux and PMIC mode to RSB will confuse Linux, which
expects the PMIC on I2C.

This explains why H6 was an exception to the pinmux value rule: the rule
was correct, and there was actually no exception. This suggests any SoC
with I2C on PL0/PL1 function 3 actually has an RSB controller present,
even if it is undocumented.

Signed-off-by: Samuel Holland <samuel@sholland.org>