Kconfig: Do not suspend OSC24M with DRAM running
Currently it is possible to configure an H6 firmware that will fail to
resume: disabling OSC24M stops the DRAM controller, but the DRAM is not
in self-refresh, so it will lose its contents.

Prevent this scenario by requiring a functioning DRAM controller driver
before CONFIG_SUSPEND_OSC24M can be enabled.

Signed-off-by: Samuel Holland <samuel@sholland.org>