drivers: dram: Add a minimal H6 DRAM controller driver
The H6 has a new generation of DRAM controller compared to the A64/H5.
Much less information is available about it than the old generation. For
that reason, this driver only contains the suspend/resume logic that is
similar to parts of the old driver.

Notably, touching any of AC, DX, or ZQ configuration or DRAM pad hold
prevented the controller from resuming. Conveniently, the one pair of
pads that we can disable are the clock lines. So this driver is not
optimal in terms of power consumption, but it does enough that we can
completely disable the DRAM controller clocks without breaking anything.
In turn, this is just enough that H6 can now select HAVE_DRAM_SUSPEND.

Signed-off-by: Samuel Holland <samuel@sholland.org>