README.md: Update H6 support status
Now that the CSS driver is feature-complete and the DRAM driver exists,
functionality for the H6 platform has caught up with the A64/H5. The new
drivers have been tested, and there are no known issues. Still, since
the code is new, I am cautiously labeling the support status as "beta"
instead of "production".

Signed-off-by: Samuel Holland <samuel@sholland.org>