drivers: css: Add an H3 CSS driver
The H3 is a 32-bit chip, with Cortex-A7 CPUs instead of the Cortex-A53
CPUs in the other supported platforms. It has a very different CPUCFG
register layout from the 64-bit chips.

For now, add just the part needed for CPU hotplug and shallow suspend.
Full CPU subsystem suspend will be added later.

Signed-off-by: Samuel Holland <samuel@sholland.org>