drivers: css: Rename CPU_SYS_RESET to be less redundant
For registers that are known to contain a single bit, including the
register name in the field name is redundant. My preference has been to
name the field after the register name minus "_REG", especially when the
field is undocumented. Do that for the CPU_SYS_RESET register, instead
of making up a field name.

Signed-off-by: Samuel Holland <samuel@sholland.org>