platform: Add a platform for the H3 SoC
This includes adding the platform config symbol, platform headers, and
the main CCU driver; and hooking up the existing R_CCU and UART drivers.
The H3 R_CCU is the same as A83T/A64/H5; and since the H3 is pin-
compatible with the H5, they have identical UART pin configurations.

SRAM A2 is very small on the H3 -- only 32 KiB total -- so the full
second half is allocated to Crust.

Signed-off-by: Samuel Holland <samuel@sholland.org>