drivers: sunxi-rsb: Support arbitrary addresses
Linux contains a table of known hardware to runtime address mappings.
Since a peripheral's runtime address cannot be reassigned without
switching it to I2C mode and back, Crust must use the same runtime
address as Linux.

A key insight is that the hardware and runtime addresses have non-
overlapping fields in RSB_SADDR_REG, and the hardware address field is
unused by commands other than RSB_SRTA. Since the fields are part of the
same 32-bit word, it takes no more work to program both than to program
just the runtime address.

So we use the combination of the two addresses, precomposed in the
correct format, as the peripheral's ID. This actually simplifies the
driver, while allowing arbitrary values for both addresses.

Signed-off-by: Samuel Holland <samuel@sholland.org>