system: Remove system state getter functions
The Boolean getters are left over from an architecture where the main
loop always called every driver and service hook, and the individual
hooks were responsible for doing different things based on the system
state. Now, the hooks are only called in the appropriate state, so these
checks have all been removed.

The integer getter was added specifically for latency calculation to
know when the state changed. This is a special piece of instrumentation
code, and not part of the firmware architecture; it does not care about
the meaning of the integer, only that the value changed. So it does not
justify exposing the state variable to the rest of the firmware.
Instead, pass the current state as an argument to the latency monitor.

Signed-off-by: Samuel Holland <samuel@sholland.org>