debug: monitor: Accept a leading 0x when parsing numbers
To be more precise: the parser now accepts [Xx] anywhere inside a
number, and interprets them as zeroes. Since the parser only cares about
the rightmost (up to) 8 characters, and since leading zeroes do not
change the value of the number, this has the desired effect of ignoring
a leading "0x".

This change significantly improves the experience of pasting numbers at
the monitor prompt.

Signed-off-by: Samuel Holland <samuel@sholland.org>