system: Prepare for regulator function rework
The current firmware design involves two lists of regulators: one to
disable during suspend and one for shutdown. The intention is to disable
additional regulators (such as those powering the DRAM) during shutdown.

That is not currently implemented. One reason for that is that the code
assumes all regulators come from the same provider. This is likely to be
the case on boards with a PMIC, but not on other boards. Most H3/H5
boards need a separate regulator driver for the CPU power domain vs
other power domains, because the CPU supply has additional control
inputs (or an entirely different control method) for DVFS support.

Un-overengineer the current code in recognition that the CPU supply is
the only supported regulator at the moment.

Signed-off-by: Samuel Holland <samuel@sholland.org>