drivers: cir: Add an RC6 protocol decoder
Use a basic state machine optimized for code size. All processing is
done in units of the sample clock, and the exact durations are provided
by the calling driver to avoid multiplication at runtime. The decoder
restarts upon encountering any error.

This decoder was written with the sunxi-cir hardware in mind, and that
is reflected in the data types in the context structure. It assumes a
relatively low sample rate, so even the longest pulses only result in a
single sample from the hardware.

Currently, only the "RC6-6A-32" variant is implemented. The trailer
(mode 0 toggle bit) is completely ignored. The decoder expects 32 data
bits and masks off the toggle bit from the data word.

Signed-off-by: Samuel Holland <samuel@sholland.org>