drivers: ccu: Split updates to PLL_CTRL_REG1
Depending on which pads the 24 MHz crystal is connected to, the
PLL_CTRL_REG1_CRYSTAL_EN bit may not have any effect. In that case,
there is no need for the extra delays or register writes.

Split the two register writes out of the helper function so one of them
can be compiled out. Currently, the helper for writing PLL_CTRL_REG1
does everything at once in an attempt to reduce code size.  In fact,
performing the two writes separately actually produces a smaller binary.

Signed-off-by: Samuel Holland <samuel@sholland.org>