drivers: ccu: Only enable X24M when necessary
Depending on which pads the 24 MHz crystal is connected to (DCXO or
X24M), the PLL_CTRL_REG1_CRYSTAL_EN bit may not have any effect on the
crystal. In that case, there is no need for the extra delays or register
writes. It also saves a bit of power while the board is running to leave
the X24M pads off.

This option is only applicable to SoCs with a separate DCXO in the RTC
domain, such as the H6. Since all H6 boards appear to use the DCXO,
select it by default there. Everywhere else, the code remains unchanged.

Signed-off-by: Samuel Holland <samuel@sholland.org>