drivers: irq: Drop IRQ accessor functions
These functions do not provide the appropriate API for use in selecting
a suspend depth. The common code does not care about the state of a
specific IRQ, but instead if any enabled peripheral/IRQ requires a
specific resource.

These functions were unused, so this does not impact the generated code.

Signed-off-by: Samuel Holland <samuel@sholland.org>