platform: Fix the name of the VDD_SYS AVCC gating field
The assumption that this bit gated VDD_SYS to VCC_PLL came mostly from
the A83T manual, which is the most recent manual to document the PRCM.
However, the A64/H3/H5 PRCM is closer to that in the A23/A31/A33, which
all have built-in audio and an RTC (unlike the A83T, which requires a
companion chip).

For the older chips, this bit is documented, and is named AVCC_A_GATING.
While AVCC is related to VCC-PLL, as it powers various analog biases, it
also powers the (analog) audio codec and the LRADC (aka KEYADC). This is
explicitly documented in the H5 datasheet. The behavior seen on the A64
matches this description: when this VDD_SYS gating bit is set, LRADC
interrupts are not received.

On the A83T, there is nothing at bit 2 of VDD_SYS_PWROFF_GATING_REG. The
only bits used by the vendor blob are the ones documented in the manual.
The PLLs (with gating) and LRADC (no gating) have separate power inputs,
and the audio codec is in a separate chip.

Signed-off-by: Samuel Holland <samuel@sholland.org>