system: Factor out the suspend depth selection
Ideally, the suspend depth (the amount of SoC that is turned off) would
be selected at runtime based on the wakeup source configuration, instead
of being fixed at build time. This allows a single binary to be shared
by various users of a board with differing wakeup needs. As a first
step, factor out the process for selecting the suspend depth, instead of
selecting it inline in the driver.

This change splits the VDD_SYS_PWROFF_GATING_REG writes during suspend,
and unconditionally performs the resume actions, but neither change
should have a functional impact.

Signed-off-by: Samuel Holland <samuel@sholland.org>