system: Automatically determine if VDD_SYS can be gated
Currently the only wakeup source that requires VDD_SYS to be ungated is
the main PIO controller. This applies if it is being polled, or if its
IRQs have been plumbed through to R_INTC.

Since using the IRQs via R_INTC is now possible, polling EINTs is now
deprecated. Keep the EINT polling code around for a while to support
older kernels that do not have the updated R_INTC driver.

Signed-off-by: Samuel Holland <samuel@sholland.org>