drivers: irq: Refactor generic/platform driver split
The EINT polling is not part of the R_INTC driver (and it will soon be
deprecated), so it stays in the generic file.

The R_INTC driver is split out and gains support for the second set of
registers found on the H6. Since that is the only platform-specific
difference, and it is only made platform-specific as a code size
optimization, keep all platforms in one file.

Signed-off-by: Samuel Holland <samuel@sholland.org>