system: Reset VDD-SYS during shutdown
Measurements show that enabling VDD_CPUS_GATING is not helpful for
reducing power consumption; in fact, SD_VDD_SYS generally doubles power
compared to SD_AVCC. However, there is one good reason to keep
VDD_CPUS_GATING: it is a prerequisite for removing power to the VDD-SYS
domain as a whole. Another requirement is to put the VDD-SYS domain in
reset. When doing both of these, it is possible to disable the regulator
supplying VDD-SYS, and that is a win for power consumption.

Resetting VDD-SYS resets the CCU, and therefore the AHB/APB domains for
all of the peripherals in the SoC. Since Linux cannot handle the CCU,
PIO, etc. being reset, this depth is only appropriate for SS_SHUTDOWN.

Signed-off-by: Samuel Holland <samuel@sholland.org>