device: Allow calling device_active() on NULL
Unlike most functions, the device_* functions allow passing in NULL, to
gracefully handle optional devices. device_get() returning an error
for NULL is no different from it returning an error because the probe
function failed. Or in other words, successfully getting a reference to
a device ensures it is nonnull.

Since device_active() is meant to be called on possibly non-active
devices, which may be NULL, handle that case without crashing.

Signed-off-by: Samuel Holland <samuel@sholland.org>