drivers: sy8106a: Remove unused regulator ID
The SY8106A has a single output, so consumers do not need to specify an
ID. The value provided in the handle is ignored by the driver. The type
definition and its use are left over from when the device and ID were
separate function parameters instead of being inside a handle; at that
point, the ID could not be left uninitialized.

Signed-off-by: Samuel Holland <samuel@sholland.org>